switch from double to float for math (no practical difference)

+++ include/mapbox/pixelmatch.hpp
@@ -10,20 +10,20 @@ namespace mapbox {
 
 namespace detail {
 
-inline double rgb2y(double r, double g, double b) {
-    return r * 0.29889531 + g * 0.58662247 + b * 0.11448223;
+inline float rgb2y(float r, float g, float b) {
+    return r * 0.29889531f + g * 0.58662247f + b * 0.11448223f;
 }
-inline double rgb2i(double r, double g, double b) {
-    return r * 0.59597799 - g * 0.27417610 - b * 0.32180189;
+inline float rgb2i(float r, float g, float b) {
+    return r * 0.59597799f - g * 0.27417610f - b * 0.32180189f;
 }
-inline double rgb2q(double r, double g, double b) {
-    return r * 0.21147017 - g * 0.52261711 + b * 0.31114694;
+inline float rgb2q(float r, float g, float b) {
+    return r * 0.21147017f - g * 0.52261711f + b * 0.31114694f;
 }
 
 // Squared YIQ distance between two pixels; semi-transparent pixels are blended against white.
 // Based on "Measuring perceived color difference using YIQ NTSC transmission color space
 // in mobile applications" by Y. Kotsarenko and F. Ramos.
-inline double colorDelta(const uint8_t* img1, const uint8_t* img2, std::size_t k, std::size_t m, bool yOnly = false) {
+inline float colorDelta(const uint8_t* img1, const uint8_t* img2, std::size_t k, std::size_t m, bool yOnly = false) {
     int r1 = img1[k], g1 = img1[k + 1], b1 = img1[k + 2], a1 = img1[k + 3];
     int r2 = img2[m], g2 = img2[m + 1], b2 = img2[m + 2], a2 = img2[m + 3];
 
@@ -31,18 +31,18 @@ inline double colorDelta(const uint8_t* img1, const uint8_t* img2, std::size_t k
 
     if (yOnly && !dri && !dgi && !dbi && !da) return 0;
 
-    double dr = dri, dg = dgi, db = dbi;
+    float dr = dri, dg = dgi, db = dbi;
     if (a1 < 255 || a2 < 255) { // blend with white background
-        dr = (r1 * a1 - r2 * a2 - 255 * da) / 255.0;
-        dg = (g1 * a1 - g2 * a2 - 255 * da) / 255.0;
-        db = (b1 * a1 - b2 * a2 - 255 * da) / 255.0;
+        dr = (r1 * a1 - r2 * a2 - 255 * da) / 255.0f;
+        dg = (g1 * a1 - g2 * a2 - 255 * da) / 255.0f;
+        db = (b1 * a1 - b2 * a2 - 255 * da) / 255.0f;
     }
 
-    double y = rgb2y(dr, dg, db);
+    float y = rgb2y(dr, dg, db);
     if (yOnly) return y;
-    double i = rgb2i(dr, dg, db);
-    double q = rgb2q(dr, dg, db);
-    return 0.5053 * y * y + 0.299 * i * i + 0.1957 * q * q;
+    float i = rgb2i(dr, dg, db);
+    float q = rgb2q(dr, dg, db);
+    return 0.5053f * y * y + 0.299f * i * i + 0.1957f * q * q;
 }
 
 inline void drawPixel(uint8_t* output, std::size_t pos, uint8_t r, uint8_t g, uint8_t b) {
@@ -54,8 +54,8 @@ inline void drawPixel(uint8_t* output, std::size_t pos, uint8_t r, uint8_t g, ui
 
 inline void drawGrayPixel(const uint8_t* img, std::size_t i, uint8_t* output, std::size_t outPos) {
     int r = img[i], g = img[i + 1], b = img[i + 2], a = img[i + 3];
-    double y = rgb2y(r, g, b);
-    uint8_t v = static_cast<uint8_t>(255 + (y - 255) * 0.1 * a / 255);
+    float y = rgb2y(r, g, b);
+    uint8_t v = static_cast<uint8_t>(255 + (y - 255) * 0.1f * a / 255);
     drawPixel(output, outPos, v, v, v);
 }
 
@@ -98,7 +98,7 @@ inline bool antialiased(const uint8_t* img,
     std::size_t y2 = std::min(y1 + 1, height - 1);
     std::size_t pos = y1 * stride + x1 * 4;
     int zeroes = (x1 == x0 || x1 == x2 || y1 == y0 || y1 == y2) ? 1 : 0;
-    double minD = 0, maxD = 0;
+    float minD = 0, maxD = 0;
     std::size_t minX = 0, minY = 0, maxX = 0, maxY = 0;
 
     // go through 8 adjacent pixels
@@ -107,7 +107,7 @@ inline bool antialiased(const uint8_t* img,
             if (x == x1 && y == y1) continue;
 
             // brightness delta between the center pixel and the adjacent one
-            double delta = colorDelta(img, img, pos, y * stride + x * 4, true);
+            float delta = colorDelta(img, img, pos, y * stride + x * 4, true);
 
             // count the number of equal, darker and brighter adjacent pixels
             if (delta == 0) {
@@ -174,7 +174,7 @@ inline uint64_t pixelmatch(const uint8_t* img1,
 
     // maximum acceptable square distance between two colors;
     // 35215 is the maximum possible value for the YIQ difference metric
-    double maxDelta = 35215 * threshold * threshold;
+    float maxDelta = 35215.0f * threshold * threshold;
     uint64_t diff = 0;
 
     // compare each pixel of one image against the other one
@@ -187,7 +187,7 @@ inline uint64_t pixelmatch(const uint8_t* img1,
 
             // squared YIQ distance between colors at this pixel position; skip the math
             // entirely if the raw RGBA bytes match
-            double delta = std::memcmp(img1 + pos1, img2 + pos2, 4) == 0 ? 0 : colorDelta(img1, img2, pos1, pos2);
+            float delta = std::memcmp(img1 + pos1, img2 + pos2, 4) == 0 ? 0 : colorDelta(img1, img2, pos1, pos2);
 
             // the color difference is above the threshold
             if (delta > maxDelta) {
